autotest/requirements.txt
requests>=2.26.0
google-api-python-client>=2.27.0
google-cloud-storage>=1.43.0